Worked on Rooms, and general properties updated.
Added a staircase room and a staircase scene. Staircase scene has an area that displays an arrow to show the player can go up. Added an arrow image. Updated properties to make the room more concise with the wall sizes and floor. Updated the camera vignette to reflect the updated room size same with Camera movement in world.